extract seed-phrase crate
