+++ Cargo.toml
@@ -140,6 +140,7 @@ members = [
     "sdk/rent",
     "sdk/sanitize",
     "sdk/seed-derivable",
+    "sdk/seed-phrase",
     "sdk/serde-varint",
     "sdk/serialize-utils",
     "sdk/sha256-hasher",
@@ -488,6 +489,7 @@ solana-remote-wallet = { path = "remote-wallet", version = "=2.2.0", default-fea
 solana-rent = { path = "sdk/rent", version = "=2.2.0", default-features = false }
 solana-sanitize = { path = "sdk/sanitize", version = "=2.2.0" }
 solana-seed-derivable = { path = "sdk/seed-derivable", version = "=2.2.0" }
+solana-seed-phrase = { path = "sdk/seed-phrase", version = "=2.2.0" }
 solana-serde-varint = { path = "sdk/serde-varint", version = "=2.2.0" }
 solana-serialize-utils = { path = "sdk/serialize-utils", version = "=2.2.0" }
 solana-sha256-hasher = { path = "sdk/sha256-hasher", version = "=2.2.0" }
