+++ sdk/Cargo.toml
@@ -37,6 +37,7 @@ full = [
     "solana-pubkey/rand",
     "dep:solana-precompile-error",
     "dep:solana-seed-derivable",
+    "dep:solana-seed-phrase",
     "dep:solana-transaction-error"
 ]
 borsh = ["dep:borsh", "solana-program/borsh", "solana-secp256k1-recover/borsh"]
@@ -116,6 +117,7 @@ solana-secp256k1-recover = { workspace = true }
 solana-serde-varint = { workspace = true }
 solana-short-vec = { workspace = true }
 solana-seed-derivable = { workspace = true, optional = true }
+solana-seed-phrase = { workspace = true, optional = true }
 solana-signer = { workspace = true, optional = true }
 solana-signature = { workspace = true, features = [
     "rand",
