+++ sdk/seed-phrase/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "solana-seed-phrase"
+description = "Solana functions for generating keypairs from seed phrases."
+documentation = "https://docs.rs/solana-seed-phrase"
+version = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+pbkdf2 = { workspace = true }
+hmac = { workspace = true }
+sha2 = { workspace = true }
+solana-signer = { workspace = true }
+
+[dev-dependencies]
+tiny-bip39 = { workspace = true }
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
