+++ sdk/signer/Cargo.toml
@@ -12,12 +12,9 @@ edition = { workspace = true }
 [dependencies]
 bs58 = { workspace = true, features = ["std"] }
 ed25519-dalek = { workspace = true }
-hmac = { workspace = true }
 itertools = { workspace = true }
-pbkdf2 = { workspace = true }
 rand0-7 = { workspace = true }
 serde_json = { workspace = true }
-sha2 = { workspace = true }
 solana-pubkey = { workspace = true }
 solana-signature = { workspace = true, features = ["verify"] }
 solana-transaction-error = { workspace = true }
@@ -25,7 +22,6 @@ thiserror = { workspace = true }
 
 [dev-dependencies]
 static_assertions = { workspace = true }
-tiny-bip39 = { workspace = true }
 
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]
