+++ sdk/signer/src/keypair.rs
@@ -3,7 +3,6 @@ use wasm_bindgen::prelude::*;
 use {
     crate::{EncodableKey, EncodableKeypair, Signer, SignerError},
     ed25519_dalek::Signer as DalekSigner,
-    hmac::Hmac,
     rand0_7::{rngs::OsRng, CryptoRng, RngCore},
     solana_pubkey::Pubkey,
     solana_signature::Signature,
@@ -198,40 +197,10 @@ pub fn keypair_from_seed(seed: &[u8]) -> Result<Keypair, Box<dyn error::Error>> 
     Ok(Keypair(dalek_keypair))
 }
 
-pub fn generate_seed_from_seed_phrase_and_passphrase(
-    seed_phrase: &str,
-    passphrase: &str,
-) -> Vec<u8> {
-    const PBKDF2_ROUNDS: u32 = 2048;
-    const PBKDF2_BYTES: usize = 64;
-
-    let salt = format!("mnemonic{passphrase}");
-
-    let mut seed = vec![0u8; PBKDF2_BYTES];
-    pbkdf2::pbkdf2::<Hmac<sha2::Sha512>>(
-        seed_phrase.as_bytes(),
-        salt.as_bytes(),
-        PBKDF2_ROUNDS,
-        &mut seed,
-    );
-    seed
-}
-
-pub fn keypair_from_seed_phrase_and_passphrase(
-    seed_phrase: &str,
-    passphrase: &str,
-) -> Result<Keypair, Box<dyn error::Error>> {
-    keypair_from_seed(&generate_seed_from_seed_phrase_and_passphrase(
-        seed_phrase,
-        passphrase,
-    ))
-}
-
 #[cfg(test)]
 mod tests {
     use {
         super::*,
-        bip39::{Language, Mnemonic, MnemonicType, Seed},
         std::{
             fs::{self, File},
             mem,
@@ -312,17 +281,6 @@ mod tests {
 
         let too_short_seed = vec![0; 31];
         assert!(keypair_from_seed(&too_short_seed).is_err());
-    }
-
-    #[test]
-    fn test_keypair_from_seed_phrase_and_passphrase() {
-        let mnemonic = Mnemonic::new(MnemonicType::Words12, Language::English);
-        let passphrase = "42";
-        let seed = Seed::new(&mnemonic, passphrase);
-        let expected_keypair = keypair_from_seed(seed.as_bytes()).unwrap();
-        let keypair =
-            keypair_from_seed_phrase_and_passphrase(mnemonic.phrase(), passphrase).unwrap();
-        assert_eq!(keypair.pubkey(), expected_keypair.pubkey());
     }
 
     #[test]
