+++ sdk/src/signer/keypair.rs
@@ -1,1 +1,7 @@
-pub use {solana_seed_derivable::keypair_from_seed_and_derivation_path, solana_signer::keypair::*};
+pub use {
+    solana_seed_derivable::keypair_from_seed_and_derivation_path,
+    solana_seed_phrase::{
+        generate_seed_from_seed_phrase_and_passphrase, keypair_from_seed_phrase_and_passphrase,
+    },
+    solana_signer::keypair::*,
+};
